Save route created
I created the route for saving new client info to the API. I was getting an error for a while becasue the 'client' in the models has a lower case 'c'. When I ran it through the Postman organization was coming up as Null. It was because the column name was misspelled.